Move ColoredTimeLaneModel in new module
